add Viper IRS files
this automatically installs the viper irs the bot provides one with when asking for irs in https://t.me/ViPER4AndroidFX
this just makes it easier for the user to use these files as they are included by default and dont have to be extracted and copied into scoped storage manually
the included zip file is a repack of the original without some broken files
if there are already IRS files installed, we do not copy the viper irs files again
All included IRS files of the provided archive have been extracted into the ViperIRS folder and all special characters have been removed.
This way managing the IRS files is way easier, and git does not complain about a large binary zip file in the repo.