Prevent OOM-killing
Prevent OOM-kills by the kernel by setting the oom_score_adj to its minimum.
We need to do this periodically, because android also changes this value on certain evens, eg assigning a lower score when the app is open and a higher score when the app is in the background.