enhance app installation
Instead of copying to /data/local we now install using a pipe,
which will not yield any file access issues anymore.
Additionally now the app will be upgraded if it exists already,
and will only be deactivated if it didn't exist beforehand.
This will make it possible to update at any point in time
and postpone the restart to a more convenient time.