Disable automatic permissions revoke
On API>=30 an apps permissions will be automatically revoked if the app is unused for a while.
As ViPER4Android is an app which is destined to be configured once and then work "unused" for a long time,
it makes sense to disable this feature.

Thanks to @binarynoise for finding the correct APPOPS operation name.